new folder structure and archtecture
