+++ lib/models/content.dart
