+++ lib/models/news_event_notice.dart
