+++ lib/theme/theme.dart
@@ -1,0 +1,2 @@
+export 'package:ufcat_app/theme/src/app_colors.dart';
+export 'package:ufcat_app/theme/src/app_font_style.dart';
